convert to using non-blocking socket io
See, the problem is that some whois servers (chinese ones, usually) either don't like to be hammered on,
or are just being disagreeable through policy.  Whatever the cause, the result is that we end up with
threads that are hung, waiting for input that's never going to arrive, on a tcp socket that's never
going to close on it's own.  We need to have a reasonable delay, since some whois servers just suck,
but we have to bail out at some point.